+++ src/main/java/com/devicehive/resource/impl/UserResourceImpl.java
@@ -131,6 +131,17 @@ public class UserResourceImpl implements UserResource {
      */
     @Override
     public Response deleteUser(long userId) {
+        HivePrincipal principal = (HivePrincipal) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        User currentUser = null;
+        if(principal.getUser() != null){
+            currentUser = principal.getUser();
+        }else if(principal.getKey() != null){
+            currentUser = principal.getKey().getUser();
+        }
+        if(currentUser != null && currentUser.getId().equals(userId)){
+            logger.debug("Rejected removing current user");
+            throw new HiveException(Messages.CANT_DELETE_CURRENT_USER_KEY, FORBIDDEN.getStatusCode());
+        }
         userService.deleteUser(userId);
         return ResponseFactory.response(NO_CONTENT);
     }
